Fixed password bug,
The crypto tool main.py had a password bug introduced with 1.0.6 that has been fixed. It now has two more options for selecting the encryption algorithm. The VMwareVMX class forgot to reset the config key when the encryption algorithm was changed. The size of the dictionaries is now calculated with a formula which reflects the structure of the dictionary.